chore(node): Move some code from the `npm` to the `bun` package
That code is now only used by `Bun`.
